changed conftest.py so that pytest will run

+++ test/conftest.py
@@ -222,14 +222,16 @@ def temp_dir():
     shutil.rmtree(dirpath)
 
 
-@pytest.fixture(scope="module", autouse=True)
+@pytest.fixture(scope="module", autouse=False)
 def delete_all_buckets():
     bucket_api = BucketAPI(APIClient(api_key, access_token, chain_name))
     buckets = bucket_api.list_buckets()
-    for bucket in buckets:
-        bucket_api.delete_bucket(bucket.bucket_name)
-        print("Deleted bucket: ", bucket.bucket_name)
+    if buckets:
+        for bucket in buckets:
+            bucket_api.delete_bucket(bucket.bucket_name)
+            print("Deleted bucket: ", bucket.bucket_name)
 
 
 def pytest_sessionfinish(session, exitstatus):
-    delete_all_buckets()
+    #delete_all_buckets()
+    pass
